+++ packages/@expo/cli/src/start/server/metro/instantiateMetro.ts
@@ -7,7 +7,12 @@ import MetroHmrServer, { Client as MetroHmrClient } from '@expo/metro/metro/HmrS
 import RevisionNotFoundError from '@expo/metro/metro/IncrementalBundler/RevisionNotFoundError';
 import type MetroServer from '@expo/metro/metro/Server';
 import formatBundlingError from '@expo/metro/metro/lib/formatBundlingError';
-import { loadConfig, resolveConfig, type ConfigT } from '@expo/metro/metro-config';
+import {
+  mergeConfig,
+  resolveConfig,
+  type InputConfigT,
+  type ConfigT,
+} from '@expo/metro/metro-config';
 import { Terminal } from '@expo/metro/metro-core';
 import { getDefaultConfig, type LoadOptions } from '@expo/metro-config';
 import chalk from 'chalk';
@@ -94,15 +99,21 @@ export async function loadMetroConfigAsync(
   const serverRoot = getMetroServerRoot(projectRoot);
   const terminalReporter = new MetroTerminalReporter(serverRoot, terminal);
 
-  const hasConfig = await resolveConfig(options.config, projectRoot);
-  let config: ConfigT = {
-    ...(await loadConfig(
-      { cwd: projectRoot, projectRoot, ...options },
-      // If the project does not have a metro.config.js, then we use the default config.
-      hasConfig.isEmpty ? getDefaultConfig(projectRoot) : undefined
-    )),
+  const defaultConfig = getDefaultConfig(projectRoot);
+  const resolvedConfig = await resolveConfig(options.config, projectRoot);
+
+  let config: ConfigT = resolvedConfig.isEmpty
+    ? defaultConfig
+    : await mergeConfig(defaultConfig, resolvedConfig.config);
+  // Set the watchfolders to include the projectRoot, as Metro assumes this
+  // Force-override the reporter
+  config = {
+    ...config,
+    watchFolders: !config.watchFolders.includes(config.projectRoot)
+      ? [config.projectRoot, ...config.watchFolders]
+      : config.watchFolders,
     reporter: {
-      update(event: any) {
+      update(event) {
         terminalReporter.update(event);
         if (reportEvent) {
           reportEvent(event);