+++ packages/@expo/metro-config/src/ExpoMetroConfig.ts
@@ -11,7 +11,7 @@ import type {
   Options as GraphOptions,
 } from '@expo/metro/metro/DeltaBundler/types';
 import { stableHash } from '@expo/metro/metro-cache';
-import type { ConfigT as MetroConfig, InputConfigT } from '@expo/metro/metro-config';
+import type { InputConfigT, ConfigT as MetroConfig } from '@expo/metro/metro-config';
 import chalk from 'chalk';
 import os from 'os';
 import path from 'path';
@@ -141,6 +141,10 @@ function memoize<T extends (...args: any[]) => any>(fn: T): T {
   }) as T;
 }
 
+function asMetroConfigInput<T extends InputConfigT>(config: T): T {
+  return config;
+}
+
 export function createStableModuleIdFactory(
   root: string
 ): (path: string, context?: { platform: string; environment?: string }) => number {
@@ -187,7 +191,7 @@ export function createStableModuleIdFactory(
 export function getDefaultConfig(
   projectRoot: string,
   { mode, isCSSEnabled = true, unstable_beforeAssetSerializationPlugins }: DefaultConfigOptions = {}
-): InputConfigT {
+) {
   const {
     getDefaultConfig: getDefaultMetroConfig,
     mergeConfig,
@@ -270,12 +274,7 @@ export function getDefaultConfig(
     console.log();
   }
 
-  const {
-    // Remove the default reporter which metro always resolves to be the react-native-community/cli reporter.
-    // This prints a giant React logo which is less accessible to users on smaller terminals.
-    reporter,
-    ...metroDefaultValues
-  } = getDefaultMetroConfig.getDefaultValues(projectRoot);
+  const metroDefaultValues = getDefaultMetroConfig.getDefaultValues(projectRoot);
 
   const cacheStore = new FileStore<any>({
     root: path.join(os.tmpdir(), 'metro-cache'),
@@ -284,9 +283,15 @@ export function getDefaultConfig(
   const serverRoot = getMetroServerRoot(projectRoot);
 
   const routerPackageRoot = resolveFrom.silent(projectRoot, 'expo-router');
-  // Merge in the default config from Metro here, even though loadConfig uses it as defaults.
-  // This is a convenience for getDefaultConfig use in metro.config.js, e.g. to modify assetExts.
-  const metroConfig: Partial<MetroConfig> = mergeConfig(metroDefaultValues, {
+
+  const expoMetroConfig = asMetroConfigInput({
+    reporter: {
+      // Remove the default reporter which metro always resolves to be the react-native-community/cli reporter.
+      // This prints a giant React logo which is less accessible to users on smaller terminals.
+      update() {
+        /*noop*/
+      },
+    },
     watchFolders,
     resolver: {
       unstable_conditionsByPlatform: {
@@ -384,7 +389,6 @@ export function getDefaultConfig(
     transformer: {
       // Custom: These are passed to `getCacheKey` and ensure invalidation when the version changes.
       unstable_renameRequire: false,
-      // @ts-expect-error: not on type.
       _expoRouterPath: routerPackageRoot ? path.relative(serverRoot, routerPackageRoot) : undefined,
       postcssHash: getPostcssConfigHash(projectRoot),
       browserslistHash: pkg?.browserslist
@@ -417,6 +421,16 @@ export function getDefaultConfig(
       }),
     },
   });
+
+  // Merge in the default config from Metro here, even though loadConfig uses it as defaults.
+  // This is a convenience for getDefaultConfig use in metro.config.js, e.g. to modify assetExts.
+  const metroConfig = mergeConfig(
+    // NOTE(@kitten): We neither want ConfigT/MetroConfig here, which is mostly marked as readonly,
+    // nor InputConfigT which is inexact and partial. Instead, we want an exact type combination of
+    // the default config and Expo's config
+    metroDefaultValues as MetroConfig & typeof expoMetroConfig,
+    expoMetroConfig
+  );
 
   return withExpoSerializers(metroConfig, { unstable_beforeAssetSerializationPlugins });
 }